+++ src/index.js
@@ -2,10 +2,13 @@ import React from "react";
 import ReactDOM from "react-dom";
 import "./index.sass";
 import App from "./App";
+import { CardContextProvider } from "./components/CardContext/CardContext";
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <CardContextProvider>
+      <App />
+    </CardContextProvider>
   </React.StrictMode>,
   document.getElementById("root")
 );
